RISC-V: Move phi-value to value file
