+++ src/riscv/lib/src/jit/builder.rs
@@ -27,11 +27,11 @@ use super::state_access::JitStateAccess;
 use super::state_access::JsaCalls;
 use crate::instruction_context::ICB;
 use crate::instruction_context::MulHighType;
-use crate::instruction_context::PhiValue;
 use crate::instruction_context::Predicate;
 use crate::instruction_context::StoreLoadInt;
 use crate::instruction_context::arithmetic::Arithmetic;
 use crate::instruction_context::comparable::Comparable;
+use crate::instruction_context::value::PhiValue;
 use crate::jit::builder::block_state::PCUpdate;
 use crate::machine_state::ProgramCounterUpdate;
 use crate::machine_state::memory::MemoryConfig;
